Use finelizers instead of ownerRefs to check if IPClaim is in use
The previous implementation was API breaking, since it required the
user/controller owning the IPClaim to remove the ownerRef before the
IPAddress would be deleted. This commit changes it so that we check for
finalizers instead. More specifically, if there are any other finalizers
than the IPAM one, we consider the IPClaim to be still in use. In this
way we "revert" the API breaking change but still get the same
functionality.