✨ Add flags for configuring rate limits
There is a built-in rate limit for the controller when talking to the
Kubernetes API. So far it has not been possible to configure this
limit. This commit adds flags for setting both the QPS and the burst
for the rate limit. The default remains the same as before (20 QPS, 30
burst).

New flags:

--kube-api-qps
--kube-api-burst

Also adds .devcontainer to .gitignore