fix ensure-golangci-lint digests
The digests for ensure-golangci-lint are incorrect, and also the
script fails to unbound variable when checking the SHA.

This fixes "make lint", and "make test", which are no longer utilized
as CI tests, but only run locally.

Additionally, make the digest check also check the ARCH, which is the
root cause of these mismatches.