use GLOBIGNORE to ignore hack/tools/go.mod
Using GLOBIGNORE makes the code simpler, and allows us to ignore the
hack/tools/go.mod for sed as well, not just in grep. This fixes issue
that module versions are reported as incorrect, if hack/tools/go.mod
is the one with wrong version, yet the output then does now show the
hack/tools/go.mod as culprit.