reduce action permissions
Contents: write is excessive right for dependabot, as we don't allow
dependabot to auto-merge anything. It is enough for it to read contents
and create pull request.

Move the permissions from top-level to job-level.

Cleanup actions formatting to be the same for easier reading and
maintenance.

This restores the token permissions score to 10/10.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>