add workflow to approve actions on PR change
If external contributor adds a PR, and a member /ok-to-test it, GH
actions do not run until approved by an approver. This approval needs
to be given if the PR content changes, so it inconveniences the
contributor but also the approver.

This PR approves workflows, if ok-to-test label is present.

Adapted from CAPI workflow:
https://github.com/kubernetes-sigs/cluster-api/blob/main/.github/workflows/pr-gh-workflow-approve.yaml

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>