set webhook min TLS version to 1.3
Webhook is connected to by kube-apiserver, and it has supported TLS 1.3
connections since Go 1.14, over 4 years ago. By default, the connection
has been negoatiated to TLS 1.3 anyways between the client and webhook.
Turning it to TLS 1.3 should not affect anything, but it shouldn't
also break anything.

In case this is an issue for some consumer, they can always use the
--tls-min-version TLS12 to turn it back to TLS 1.2.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>