Optimize the dependabot config
This commit adds newly introduced multi-directories feature in dependabot. With this hopefully all the go modules would be bumped in one PR for a dependancy.
Also this commit optimizes the dependabot configs a bit by adding a group for k8s.io dependancy bump and allowing patch versions for the ignore rules.

Signed-off-by: Kashif Khan <kashif.khan@est.tech>