Fix issues in running local hack scripts
In IPAM, there was only issues in codegen. Fix that, plus made it
dynamically detect all the config and generated files, so they're
included if they'd be generated elsewhere. It is in line with BMO
generate.sh now.

For others, simply adding one shellcheck ignore for a optional check
many people have enabled about, tired of seeing my editor complain :)

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>