correct the tag comment to 4.0.1
3.7.1 and 4.0.0 are the same tag, so in a previous PR we corrected
the tag to 3.7.1 as dependabot was not supposed to bump minors. But
since it actually doesn't honor the comment tag, but checks the sha
and sees a new version, it thinks were using v4 already. Hence this
PR. So since 3.7.1 and 4.0.0 are the same and it works, we can take
4.0.1 as well, but let's update the comment tag to reflect reality.

Also, fix the identation.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>