Use OSV-Scanner release binary and Reporter action workflow instead of pointing to a file
Pointing to the workflow file in OSV scanner repository leads to use a commit sha which doesn't correspond to the osv scanner's release tag sha. Hence it is cumbersome to manually check and update the file's sha everytime a new release is there. This PR uses OSV scanner binary and Reported action workflow from release and the sha pinning also corresponds to the release tag. We also tried to use OSV scanner workflow instead of installing the binary but the workflow build the OSV scanner container and the container doesnt have go binary, hence go call analysis fails. Hence we install and use the binary instead.

Also changing the schedule scan time to 6 hours earlier to make it easier to debug the issue.

Signed-off-by: Kashif Khan <kashif.khan@est.tech>