lint release program
Adding release to golangci-lint workflow and makefile to bring it inline with the rest of the repository.

Addressed lint issues found by the linters.

For golangci-lint to be able to lint the hack/tools/release program, the build
tags argument and modules-download-mode of readonly must be passed.

Instead of a new job to only pass in additional argument. Github's 'includes' functionality of
matricies was used.

Signed-off-by: Wall Christopher <wall.christopher.d@gmail.com>