Add release program to golangci-lint workflow
Adding release to golangci-lint workflow to bring it inline with
the rest of the repository.

For golangci-lint to be able to lint the hack/tools/release program, the build
tags argument and modules-download-mode of readonly must be passed.

Instead of a new job to only pass in additional argument. Github's 'includes' functionality of
matricies was used.

Signed-off-by: Wall Christopher <wall.christopher.d@gmail.com>